requirements.txt
pyadomd==0.1.1
python-dotenv==1.1.0
azure-identity==1.21.0
requests==2.32.3
mcp>=1.0.0
